Move SkillService file
